Passing a singleton can be problematic when dealing with multiple threads. Because the object is shared among threads - if there is any access to an object's internal properties that are shared, we could run into trouble (e.g. race conditions). Now, for the Dialer specificlly, this technically should not be an issue. The Dialer is assumed to be simply making a network call - all we need to do is wait for a 200 response (not even a DialStatus really).
However, I've decided to split out the Dialer so that we pass a new instance on every call - this avoid the potential pitfalls of utilizing a singleton.

I've also fixed the test for 'test_power_dialer_failed_dials'. The previous test was flawed and just... wrong. This updated test now assumes that some network error occurred and we have a "FAILED" status returned.